+++ tests/Firebase/Unit/Database/Query/Sorter/OrderByValueTest.php
@@ -1,0 +1,60 @@
+<?php
+
+namespace Kreait\Tests\Firebase\Unit\Database\Query\Sorter;
+
+use GuzzleHttp\Psr7\Uri;
+use Kreait\Firebase\Database\Query\Sorter\OrderByValue;
+use Kreait\Tests\Firebase\Unit\UnitTestCase;
+
+class OrderByValueTest extends UnitTestCase
+{
+    /**
+     * @var OrderByValue
+     */
+    protected $sorter;
+
+    protected function setUp()
+    {
+        $this->sorter = new OrderByValue();
+    }
+
+    public function testModifyUri()
+    {
+        $this->assertContains(
+            'orderBy='.rawurlencode('"$value"'),
+            (string) $this->sorter->modifyUri(new Uri('http://domain.tld'))
+        );
+    }
+
+    /**
+     * @dataProvider valueProvider
+     */
+    public function testModifyValue($expected, $value)
+    {
+        $this->assertSame($expected, $this->sorter->modifyValue($value));
+    }
+
+    public function valueProvider()
+    {
+        return [
+            'scalar' => [
+                'expected' => 'scalar',
+                'given' => 'scalar',
+            ],
+            'array' => [
+                'expected' => [
+                    'third' => 1,
+                    'fourth' => 2,
+                    'first' => 3,
+                    'second' => 4,
+                ],
+                'given' => [
+                    'first' => 3,
+                    'second' => 4,
+                    'third' => 1,
+                    'fourth' => 2,
+                ],
+            ],
+        ];
+    }
+}